Apply suggestions from code review

+++ app/src/main/java/com/openlauncher/app/ui/widget/SpeedometerWidget.kt
@@ -102,7 +102,7 @@ fun SpeedometerWidget(
 
                 for (i in 0..10) {
                     val angle   = startAngle + i * (sweepTotal / 10f)
-                    val rad     = angle.toDouble() * (PI / 180.0)
+                    val rad     = angle * (PI.toFloat() / 180f)
                     val isMajor = i % 2 == 0
                     val outerR  = arcR - trackW / 2f - 3.dp.toPx()
                     val innerR  = outerR - if (isMajor) 7.dp.toPx() else 4.dp.toPx()
